updated icad_tone_detect_handler.py and fixed a bug with cleaning files and archiving files.

+++ lib/audio_file_handler.py
@@ -133,6 +133,7 @@ def clean_files(archive_path, archive_days):
     archive_dir = os.path.abspath(archive_path)
 
     for root, dirs, files in os.walk(archive_dir, topdown=False):
+        # Clean files that are older than archive_days
         for name in files:
             file_path = os.path.join(root, name)
             creation_time = os.path.getctime(file_path)
@@ -148,13 +149,16 @@ def clean_files(archive_path, archive_days):
                     traceback.print_exc()
 
         # Clean up empty directories
-
         for dir_name in dirs:
             dir_path = os.path.join(root, dir_name)
-            try:
-                os.rmdir(dir_path)
-            except Exception as e:
-                module_logger.error(f"Failed to remove directory: {dir_path}, Error: {str(e)}")
-                traceback.print_exc()
+            # Check if the directory is empty
+            if not os.listdir(dir_path):
+                try:
+                    os.rmdir(dir_path)
+                    module_logger.debug(f"Successfully removed empty directory: {dir_path}")
+                except Exception as e:
+                    module_logger.error(f"Failed to remove directory: {dir_path}, Error: {str(e)}")
+                    traceback.print_exc()
 
     module_logger.info(f"Cleaned {count} files.")
+
